@@ -19,21 +19,20 @@
         }
     ],
     "require": {
-        "php": ">=7.1",
+        "php": ">=7.2",
         "ext-curl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
         "guzzlehttp/guzzle": "^6.2 || ^7.2"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.4",
-        "squizlabs/php_codesniffer": "~2.6",
-        "friendsofphp/php-cs-fixer": "~2.12"
+        "phpunit/phpunit": "^8.0 || ^9.0",
+        "friendsofphp/php-cs-fixer": "^2.12"
     },
     "autoload": {
         "psr-4": { "Brave\\NeucoreApi\\" : "lib/" }
     },
     "autoload-dev": {
-        "psr-4": { "Brave\\NeucoreApi\\" : "test/" }
+        "psr-4": { "Brave\\NeucoreApi\\Test\\" : "test/" }
     }
 }
