@@ -1,7 +1,7 @@
 <?php
 /**
  * Configuration
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -14,10 +14,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -30,7 +29,7 @@ namespace Brave\NeucoreApi;
 
 /**
  * Configuration Class Doc Comment
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -39,6 +38,9 @@ namespace Brave\NeucoreApi;
  */
 class Configuration
 {
+    /**
+     * @var Configuration
+     */
     private static $defaultConfiguration;
 
     /**
@@ -138,7 +140,7 @@ class Configuration
      *
      * @param string $apiKeyIdentifier API key identifier (authentication scheme)
      *
-     * @return string API key or token
+     * @return null|string API key or token
      */
     public function getApiKey($apiKeyIdentifier)
     {
@@ -164,7 +166,7 @@ class Configuration
      *
      * @param string $apiKeyIdentifier API key identifier (authentication scheme)
      *
-     * @return string
+     * @return null|string
      */
     public function getApiKeyPrefix($apiKeyIdentifier)
     {
@@ -396,7 +398,7 @@ class Configuration
         $report  = 'PHP SDK (Brave\NeucoreApi) Debug Report:' . PHP_EOL;
         $report .= '    OS: ' . php_uname() . PHP_EOL;
         $report .= '    PHP Version: ' . PHP_VERSION . PHP_EOL;
-        $report .= '    The version of the OpenAPI document: 1.14.0' . PHP_EOL;
+        $report .= '    The version of the OpenAPI document: 1.21.0' . PHP_EOL;
         $report .= '    Temp Folder Path: ' . self::getDefaultConfiguration()->getTempFolderPath() . PHP_EOL;
 
         return $report;
@@ -407,7 +409,7 @@ class Configuration
      *
      * @param  string $apiKeyIdentifier name of apikey
      *
-     * @return string API key with the prefix
+     * @return null|string API key with the prefix
      */
     public function getApiKeyWithPrefix($apiKeyIdentifier)
     {
@@ -430,29 +432,29 @@ class Configuration
     /**
      * Returns an array of host settings
      *
-     * @return an array of host settings
+     * @return array an array of host settings
      */
     public function getHostSettings()
     {
-        return array(
-          array(
-            "url" => "https://localhost/api",
-            "description" => "No description provided",
-          )
-        );
+        return [
+            [
+                "url" => "https://localhost/api",
+                "description" => "No description provided",
+            ]
+        ];
     }
 
     /**
      * Returns URL based on the index and variables
      *
-     * @param index array index of the host settings
-     * @param variables hash of variable and the corresponding value (optional)
-     * @return URL based on host settings
+     * @param int        $index     index of the host settings
+     * @param array|null $variables hash of variable and the corresponding value (optional)
+     * @return string URL based on host settings
      */
     public function getHostFromSettings($index, $variables = null)
     {
         if (null === $variables) {
-            $variables = array();
+            $variables = [];
         }
 
         $hosts = $this->getHostSettings();
@@ -466,9 +468,9 @@ class Configuration
         $url = $host["url"];
 
         // go through variable and assign a value
-        foreach ($host["variables"] as $name => $variable) {
+        foreach ($host["variables"] ?? [] as $name => $variable) {
             if (array_key_exists($name, $variables)) { // check to see if it's in the variables provided by the user
-                if (in_array($variables[$name], $variable["enum_values"])) { // check to see if the value is in the enum
+                if (in_array($variables[$name], $variable["enum_values"], true)) { // check to see if the value is in the enum
                     $url = str_replace("{".$name."}", $variables[$name], $url);
                 } else {
                     throw new \InvalidArgumentException("The variable `$name` in the host URL has invalid value ".$variables[$name].". Must be ".join(',', $variable["enum_values"]).".");
