@@ -2,7 +2,7 @@
 /**
  * Alliance
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class Alliance implements ModelInterface, ArrayAccess
+class Alliance implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -67,6 +69,8 @@ class Alliance implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => 'int64',
@@ -173,9 +177,6 @@ class Alliance implements ModelInterface, ArrayAccess
         return self::$openAPIModelName;
     }
 
-    
-
-    
 
     /**
      * Associative array for storing property values
@@ -192,10 +193,10 @@ class Alliance implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['ticker'] = isset($data['ticker']) ? $data['ticker'] : null;
-        $this->container['groups'] = isset($data['groups']) ? $data['groups'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['ticker'] = $data['ticker'] ?? null;
+        $this->container['groups'] = $data['groups'] ?? null;
     }
 
     /**
@@ -246,7 +247,7 @@ class Alliance implements ModelInterface, ArrayAccess
      *
      * @param int $id EVE alliance ID.
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -270,7 +271,7 @@ class Alliance implements ModelInterface, ArrayAccess
      *
      * @param string $name EVE alliance name.
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -294,7 +295,7 @@ class Alliance implements ModelInterface, ArrayAccess
      *
      * @param string $ticker Alliance ticker.
      *
-     * @return $this
+     * @return self
      */
     public function setTicker($ticker)
     {
@@ -318,7 +319,7 @@ class Alliance implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Group[]|null $groups Groups for automatic assignment (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setGroups($groups)
     {
@@ -343,18 +344,18 @@ class Alliance implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -377,6 +378,18 @@ class Alliance implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
