@@ -2,7 +2,7 @@
 /**
  * Character
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class Character implements ModelInterface, ArrayAccess
+class Character implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -64,13 +66,16 @@ class Character implements ModelInterface, ArrayAccess
         'valid_token_time' => '\DateTime',
         'created' => '\DateTime',
         'last_update' => '\DateTime',
-        'corporation' => '\Brave\NeucoreApi\Model\Corporation'
+        'corporation' => '\Brave\NeucoreApi\Model\Corporation',
+        'character_name_changes' => '\Brave\NeucoreApi\Model\CharacterNameChange[]'
     ];
 
     /**
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => 'int64',
@@ -80,7 +85,8 @@ class Character implements ModelInterface, ArrayAccess
         'valid_token_time' => 'date-time',
         'created' => 'date-time',
         'last_update' => 'date-time',
-        'corporation' => null
+        'corporation' => null,
+        'character_name_changes' => null
     ];
 
     /**
@@ -117,7 +123,8 @@ class Character implements ModelInterface, ArrayAccess
         'valid_token_time' => 'validTokenTime',
         'created' => 'created',
         'last_update' => 'lastUpdate',
-        'corporation' => 'corporation'
+        'corporation' => 'corporation',
+        'character_name_changes' => 'characterNameChanges'
     ];
 
     /**
@@ -133,7 +140,8 @@ class Character implements ModelInterface, ArrayAccess
         'valid_token_time' => 'setValidTokenTime',
         'created' => 'setCreated',
         'last_update' => 'setLastUpdate',
-        'corporation' => 'setCorporation'
+        'corporation' => 'setCorporation',
+        'character_name_changes' => 'setCharacterNameChanges'
     ];
 
     /**
@@ -149,7 +157,8 @@ class Character implements ModelInterface, ArrayAccess
         'valid_token_time' => 'getValidTokenTime',
         'created' => 'getCreated',
         'last_update' => 'getLastUpdate',
-        'corporation' => 'getCorporation'
+        'corporation' => 'getCorporation',
+        'character_name_changes' => 'getCharacterNameChanges'
     ];
 
     /**
@@ -193,9 +202,6 @@ class Character implements ModelInterface, ArrayAccess
         return self::$openAPIModelName;
     }
 
-    
-
-    
 
     /**
      * Associative array for storing property values
@@ -212,14 +218,15 @@ class Character implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['main'] = isset($data['main']) ? $data['main'] : null;
-        $this->container['valid_token'] = isset($data['valid_token']) ? $data['valid_token'] : null;
-        $this->container['valid_token_time'] = isset($data['valid_token_time']) ? $data['valid_token_time'] : null;
-        $this->container['created'] = isset($data['created']) ? $data['created'] : null;
-        $this->container['last_update'] = isset($data['last_update']) ? $data['last_update'] : null;
-        $this->container['corporation'] = isset($data['corporation']) ? $data['corporation'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['main'] = $data['main'] ?? null;
+        $this->container['valid_token'] = $data['valid_token'] ?? null;
+        $this->container['valid_token_time'] = $data['valid_token_time'] ?? null;
+        $this->container['created'] = $data['created'] ?? null;
+        $this->container['last_update'] = $data['last_update'] ?? null;
+        $this->container['corporation'] = $data['corporation'] ?? null;
+        $this->container['character_name_changes'] = $data['character_name_changes'] ?? null;
     }
 
     /**
@@ -267,7 +274,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param int $id EVE character ID.
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -291,7 +298,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param string $name EVE character name.
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -315,7 +322,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param bool|null $main main
      *
-     * @return $this
+     * @return self
      */
     public function setMain($main)
     {
@@ -339,7 +346,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param bool|null $valid_token Shows if character's refresh token is valid or not.  This is null if there is no refresh token (EVE SSOv1 only) or a valid token but without scopes (SSOv2).
      *
-     * @return $this
+     * @return self
      */
     public function setValidToken($valid_token)
     {
@@ -363,7 +370,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $valid_token_time Date and time when that valid token property was last changed.
      *
-     * @return $this
+     * @return self
      */
     public function setValidTokenTime($valid_token_time)
     {
@@ -387,7 +394,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $created created
      *
-     * @return $this
+     * @return self
      */
     public function setCreated($created)
     {
@@ -411,7 +418,7 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $last_update Last ESI update.
      *
-     * @return $this
+     * @return self
      */
     public function setLastUpdate($last_update)
     {
@@ -435,11 +442,35 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Corporation|null $corporation corporation
      *
-     * @return $this
+     * @return self
      */
     public function setCorporation($corporation)
     {
         $this->container['corporation'] = $corporation;
+
+        return $this;
+    }
+
+    /**
+     * Gets character_name_changes
+     *
+     * @return \Brave\NeucoreApi\Model\CharacterNameChange[]|null
+     */
+    public function getCharacterNameChanges()
+    {
+        return $this->container['character_name_changes'];
+    }
+
+    /**
+     * Sets character_name_changes
+     *
+     * @param \Brave\NeucoreApi\Model\CharacterNameChange[]|null $character_name_changes List of previous character names (API: not included by default).
+     *
+     * @return self
+     */
+    public function setCharacterNameChanges($character_name_changes)
+    {
+        $this->container['character_name_changes'] = $character_name_changes;
 
         return $this;
     }
@@ -460,18 +491,18 @@ class Character implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -494,6 +525,18 @@ class Character implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
