@@ -2,7 +2,7 @@
 /**
  * Corporation
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class Corporation implements ModelInterface, ArrayAccess
+class Corporation implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -70,6 +72,8 @@ class Corporation implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => 'int64',
@@ -188,9 +192,6 @@ class Corporation implements ModelInterface, ArrayAccess
         return self::$openAPIModelName;
     }
 
-    
-
-    
 
     /**
      * Associative array for storing property values
@@ -207,13 +208,13 @@ class Corporation implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['ticker'] = isset($data['ticker']) ? $data['ticker'] : null;
-        $this->container['alliance'] = isset($data['alliance']) ? $data['alliance'] : null;
-        $this->container['groups'] = isset($data['groups']) ? $data['groups'] : null;
-        $this->container['tracking_last_update'] = isset($data['tracking_last_update']) ? $data['tracking_last_update'] : null;
-        $this->container['auto_allowlist'] = isset($data['auto_allowlist']) ? $data['auto_allowlist'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['ticker'] = $data['ticker'] ?? null;
+        $this->container['alliance'] = $data['alliance'] ?? null;
+        $this->container['groups'] = $data['groups'] ?? null;
+        $this->container['tracking_last_update'] = $data['tracking_last_update'] ?? null;
+        $this->container['auto_allowlist'] = $data['auto_allowlist'] ?? null;
     }
 
     /**
@@ -264,7 +265,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param int $id EVE corporation ID.
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -288,7 +289,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param string $name EVE corporation name.
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -312,7 +313,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param string $ticker Corporation ticker.
      *
-     * @return $this
+     * @return self
      */
     public function setTicker($ticker)
     {
@@ -336,7 +337,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Alliance|null $alliance alliance
      *
-     * @return $this
+     * @return self
      */
     public function setAlliance($alliance)
     {
@@ -360,7 +361,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Group[]|null $groups Groups for automatic assignment (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setGroups($groups)
     {
@@ -384,7 +385,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $tracking_last_update Last update of corporation member tracking data (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setTrackingLastUpdate($tracking_last_update)
     {
@@ -408,7 +409,7 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param bool|null $auto_allowlist True if this corporation was automatically placed on the allowlist of a watchlist (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setAutoAllowlist($auto_allowlist)
     {
@@ -433,18 +434,18 @@ class Corporation implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -467,6 +468,18 @@ class Corporation implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
