@@ -2,7 +2,7 @@
 /**
  * CorporationMember
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -40,10 +39,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class CorporationMember implements ModelInterface, ArrayAccess
+class CorporationMember implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -76,6 +78,8 @@ class CorporationMember implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'player' => null,
@@ -214,9 +218,6 @@ class CorporationMember implements ModelInterface, ArrayAccess
         return self::$openAPIModelName;
     }
 
-    
-
-    
 
     /**
      * Associative array for storing property values
@@ -233,18 +234,18 @@ class CorporationMember implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['player'] = isset($data['player']) ? $data['player'] : null;
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['location'] = isset($data['location']) ? $data['location'] : null;
-        $this->container['logoff_date'] = isset($data['logoff_date']) ? $data['logoff_date'] : null;
-        $this->container['logon_date'] = isset($data['logon_date']) ? $data['logon_date'] : null;
-        $this->container['ship_type'] = isset($data['ship_type']) ? $data['ship_type'] : null;
-        $this->container['start_date'] = isset($data['start_date']) ? $data['start_date'] : null;
-        $this->container['character'] = isset($data['character']) ? $data['character'] : null;
-        $this->container['missing_character_mail_sent_date'] = isset($data['missing_character_mail_sent_date']) ? $data['missing_character_mail_sent_date'] : null;
-        $this->container['missing_character_mail_sent_result'] = isset($data['missing_character_mail_sent_result']) ? $data['missing_character_mail_sent_result'] : null;
-        $this->container['missing_character_mail_sent_number'] = isset($data['missing_character_mail_sent_number']) ? $data['missing_character_mail_sent_number'] : null;
+        $this->container['player'] = $data['player'] ?? null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['location'] = $data['location'] ?? null;
+        $this->container['logoff_date'] = $data['logoff_date'] ?? null;
+        $this->container['logon_date'] = $data['logon_date'] ?? null;
+        $this->container['ship_type'] = $data['ship_type'] ?? null;
+        $this->container['start_date'] = $data['start_date'] ?? null;
+        $this->container['character'] = $data['character'] ?? null;
+        $this->container['missing_character_mail_sent_date'] = $data['missing_character_mail_sent_date'] ?? null;
+        $this->container['missing_character_mail_sent_result'] = $data['missing_character_mail_sent_result'] ?? null;
+        $this->container['missing_character_mail_sent_number'] = $data['missing_character_mail_sent_number'] ?? null;
     }
 
     /**
@@ -292,7 +293,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Player|null $player player
      *
-     * @return $this
+     * @return self
      */
     public function setPlayer($player)
     {
@@ -316,7 +317,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param int $id EVE Character ID.
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -340,7 +341,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param string $name EVE Character name.
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -364,7 +365,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\EsiLocation|null $location location
      *
-     * @return $this
+     * @return self
      */
     public function setLocation($location)
     {
@@ -388,7 +389,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $logoff_date logoff_date
      *
-     * @return $this
+     * @return self
      */
     public function setLogoffDate($logoff_date)
     {
@@ -412,7 +413,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $logon_date logon_date
      *
-     * @return $this
+     * @return self
      */
     public function setLogonDate($logon_date)
     {
@@ -436,7 +437,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\EsiType|null $ship_type ship_type
      *
-     * @return $this
+     * @return self
      */
     public function setShipType($ship_type)
     {
@@ -460,7 +461,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $start_date start_date
      *
-     * @return $this
+     * @return self
      */
     public function setStartDate($start_date)
     {
@@ -484,7 +485,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Character|null $character character
      *
-     * @return $this
+     * @return self
      */
     public function setCharacter($character)
     {
@@ -508,7 +509,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param \DateTime|null $missing_character_mail_sent_date Date and time of the last sent mail.
      *
-     * @return $this
+     * @return self
      */
     public function setMissingCharacterMailSentDate($missing_character_mail_sent_date)
     {
@@ -532,7 +533,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param string|null $missing_character_mail_sent_result Result of the last sent mail (OK, Blocked, CSPA charge > 0)
      *
-     * @return $this
+     * @return self
      */
     public function setMissingCharacterMailSentResult($missing_character_mail_sent_result)
     {
@@ -556,7 +557,7 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param int|null $missing_character_mail_sent_number Number of mails sent, is reset when the character is added.
      *
-     * @return $this
+     * @return self
      */
     public function setMissingCharacterMailSentNumber($missing_character_mail_sent_number)
     {
@@ -581,18 +582,18 @@ class CorporationMember implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -615,6 +616,18 @@ class CorporationMember implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
