@@ -2,7 +2,7 @@
 /**
  * EsiLocation
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class EsiLocation implements ModelInterface, ArrayAccess
+class EsiLocation implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -66,6 +68,8 @@ class EsiLocation implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => 'int64',
@@ -171,9 +175,7 @@ class EsiLocation implements ModelInterface, ArrayAccess
     const CATEGORY_SYSTEM = 'system';
     const CATEGORY_STATION = 'station';
     const CATEGORY_STRUCTURE = 'structure';
-    
 
-    
     /**
      * Gets allowable values of the enum
      *
@@ -187,7 +189,6 @@ class EsiLocation implements ModelInterface, ArrayAccess
             self::CATEGORY_STRUCTURE,
         ];
     }
-    
 
     /**
      * Associative array for storing property values
@@ -204,9 +205,9 @@ class EsiLocation implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['category'] = isset($data['category']) ? $data['category'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['category'] = $data['category'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
     }
 
     /**
@@ -227,7 +228,8 @@ class EsiLocation implements ModelInterface, ArrayAccess
         $allowedValues = $this->getCategoryAllowableValues();
         if (!is_null($this->container['category']) && !in_array($this->container['category'], $allowedValues, true)) {
             $invalidProperties[] = sprintf(
-                "invalid value for 'category', must be one of '%s'",
+                "invalid value '%s' for 'category', must be one of '%s'",
+                $this->container['category'],
                 implode("', '", $allowedValues)
             );
         }
@@ -265,7 +267,7 @@ class EsiLocation implements ModelInterface, ArrayAccess
      *
      * @param int $id id
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -289,7 +291,7 @@ class EsiLocation implements ModelInterface, ArrayAccess
      *
      * @param string $category category
      *
-     * @return $this
+     * @return self
      */
     public function setCategory($category)
     {
@@ -297,7 +299,8 @@ class EsiLocation implements ModelInterface, ArrayAccess
         if (!in_array($category, $allowedValues, true)) {
             throw new \InvalidArgumentException(
                 sprintf(
-                    "Invalid value for 'category', must be one of '%s'",
+                    "Invalid value '%s' for 'category', must be one of '%s'",
+                    $category,
                     implode("', '", $allowedValues)
                 )
             );
@@ -322,7 +325,7 @@ class EsiLocation implements ModelInterface, ArrayAccess
      *
      * @param string $name name
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -347,18 +350,18 @@ class EsiLocation implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -381,6 +384,18 @@ class EsiLocation implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
