@@ -2,7 +2,7 @@
 /**
  * Group
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class Group implements ModelInterface, ArrayAccess
+class Group implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -66,6 +68,8 @@ class Group implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => null,
@@ -171,9 +175,7 @@ class Group implements ModelInterface, ArrayAccess
     const VISIBILITY__PRIVATE = 'private';
     const VISIBILITY__PUBLIC = 'public';
     const VISIBILITY_CONDITIONED = 'conditioned';
-    
 
-    
     /**
      * Gets allowable values of the enum
      *
@@ -187,7 +189,6 @@ class Group implements ModelInterface, ArrayAccess
             self::VISIBILITY_CONDITIONED,
         ];
     }
-    
 
     /**
      * Associative array for storing property values
@@ -204,9 +205,9 @@ class Group implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['visibility'] = isset($data['visibility']) ? $data['visibility'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['visibility'] = $data['visibility'] ?? null;
     }
 
     /**
@@ -235,7 +236,8 @@ class Group implements ModelInterface, ArrayAccess
         $allowedValues = $this->getVisibilityAllowableValues();
         if (!is_null($this->container['visibility']) && !in_array($this->container['visibility'], $allowedValues, true)) {
             $invalidProperties[] = sprintf(
-                "invalid value for 'visibility', must be one of '%s'",
+                "invalid value '%s' for 'visibility', must be one of '%s'",
+                $this->container['visibility'],
                 implode("', '", $allowedValues)
             );
         }
@@ -270,7 +272,7 @@ class Group implements ModelInterface, ArrayAccess
      *
      * @param int $id Group ID.
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -294,7 +296,7 @@ class Group implements ModelInterface, ArrayAccess
      *
      * @param string $name A unique group name (can be changed).
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -325,7 +327,7 @@ class Group implements ModelInterface, ArrayAccess
      *
      * @param string|null $visibility visibility
      *
-     * @return $this
+     * @return self
      */
     public function setVisibility($visibility)
     {
@@ -333,7 +335,8 @@ class Group implements ModelInterface, ArrayAccess
         if (!is_null($visibility) && !in_array($visibility, $allowedValues, true)) {
             throw new \InvalidArgumentException(
                 sprintf(
-                    "Invalid value for 'visibility', must be one of '%s'",
+                    "Invalid value '%s' for 'visibility', must be one of '%s'",
+                    $visibility,
                     implode("', '", $allowedValues)
                 )
             );
@@ -359,18 +362,18 @@ class Group implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -393,6 +396,18 @@ class Group implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
