@@ -2,7 +2,7 @@
 /**
  * GroupApplication
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -40,10 +39,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class GroupApplication implements ModelInterface, ArrayAccess
+class GroupApplication implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -69,6 +71,8 @@ class GroupApplication implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'id' => null,
@@ -182,9 +186,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
     const STATUS_PENDING = 'pending';
     const STATUS_ACCEPTED = 'accepted';
     const STATUS_DENIED = 'denied';
-    
 
-    
     /**
      * Gets allowable values of the enum
      *
@@ -198,7 +200,6 @@ class GroupApplication implements ModelInterface, ArrayAccess
             self::STATUS_DENIED,
         ];
     }
-    
 
     /**
      * Associative array for storing property values
@@ -215,11 +216,11 @@ class GroupApplication implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['player'] = isset($data['player']) ? $data['player'] : null;
-        $this->container['group'] = isset($data['group']) ? $data['group'] : null;
-        $this->container['created'] = isset($data['created']) ? $data['created'] : null;
-        $this->container['status'] = isset($data['status']) ? $data['status'] : null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['player'] = $data['player'] ?? null;
+        $this->container['group'] = $data['group'] ?? null;
+        $this->container['created'] = $data['created'] ?? null;
+        $this->container['status'] = $data['status'] ?? null;
     }
 
     /**
@@ -246,7 +247,8 @@ class GroupApplication implements ModelInterface, ArrayAccess
         $allowedValues = $this->getStatusAllowableValues();
         if (!is_null($this->container['status']) && !in_array($this->container['status'], $allowedValues, true)) {
             $invalidProperties[] = sprintf(
-                "invalid value for 'status', must be one of '%s'",
+                "invalid value '%s' for 'status', must be one of '%s'",
+                $this->container['status'],
                 implode("', '", $allowedValues)
             );
         }
@@ -281,7 +283,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param int $id id
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -305,7 +307,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Player $player player
      *
-     * @return $this
+     * @return self
      */
     public function setPlayer($player)
     {
@@ -329,7 +331,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Group $group group
      *
-     * @return $this
+     * @return self
      */
     public function setGroup($group)
     {
@@ -353,7 +355,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param \DateTime $created created
      *
-     * @return $this
+     * @return self
      */
     public function setCreated($created)
     {
@@ -377,7 +379,7 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param string|null $status Group application status.
      *
-     * @return $this
+     * @return self
      */
     public function setStatus($status)
     {
@@ -385,7 +387,8 @@ class GroupApplication implements ModelInterface, ArrayAccess
         if (!is_null($status) && !in_array($status, $allowedValues, true)) {
             throw new \InvalidArgumentException(
                 sprintf(
-                    "Invalid value for 'status', must be one of '%s'",
+                    "Invalid value '%s' for 'status', must be one of '%s'",
+                    $status,
                     implode("', '", $allowedValues)
                 )
             );
@@ -411,18 +414,18 @@ class GroupApplication implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -445,6 +448,18 @@ class GroupApplication implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
