@@ -2,7 +2,7 @@
 /**
  * Player
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class Player implements ModelInterface, ArrayAccess
+class Player implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -57,6 +59,7 @@ class Player implements ModelInterface, ArrayAccess
       * @var string[]
       */
     protected static $openAPITypes = [
+        'service_accounts' => '\Brave\NeucoreApi\Model\ServiceAccount[]',
         'id' => 'int',
         'name' => 'string',
         'status' => 'string',
@@ -73,8 +76,11 @@ class Player implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
+        'service_accounts' => null,
         'id' => null,
         'name' => null,
         'status' => null,
@@ -114,6 +120,7 @@ class Player implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $attributeMap = [
+        'service_accounts' => 'serviceAccounts',
         'id' => 'id',
         'name' => 'name',
         'status' => 'status',
@@ -132,6 +139,7 @@ class Player implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $setters = [
+        'service_accounts' => 'setServiceAccounts',
         'id' => 'setId',
         'name' => 'setName',
         'status' => 'setStatus',
@@ -150,6 +158,7 @@ class Player implements ModelInterface, ArrayAccess
      * @var string[]
      */
     protected static $getters = [
+        'service_accounts' => 'getServiceAccounts',
         'id' => 'getId',
         'name' => 'getName',
         'status' => 'getStatus',
@@ -205,9 +214,7 @@ class Player implements ModelInterface, ArrayAccess
 
     const STATUS_STANDARD = 'standard';
     const STATUS_MANAGED = 'managed';
-    
 
-    
     /**
      * Gets allowable values of the enum
      *
@@ -220,7 +227,6 @@ class Player implements ModelInterface, ArrayAccess
             self::STATUS_MANAGED,
         ];
     }
-    
 
     /**
      * Associative array for storing property values
@@ -237,16 +243,17 @@ class Player implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['id'] = isset($data['id']) ? $data['id'] : null;
-        $this->container['name'] = isset($data['name']) ? $data['name'] : null;
-        $this->container['status'] = isset($data['status']) ? $data['status'] : null;
-        $this->container['roles'] = isset($data['roles']) ? $data['roles'] : null;
-        $this->container['characters'] = isset($data['characters']) ? $data['characters'] : null;
-        $this->container['groups'] = isset($data['groups']) ? $data['groups'] : null;
-        $this->container['manager_groups'] = isset($data['manager_groups']) ? $data['manager_groups'] : null;
-        $this->container['manager_apps'] = isset($data['manager_apps']) ? $data['manager_apps'] : null;
-        $this->container['removed_characters'] = isset($data['removed_characters']) ? $data['removed_characters'] : null;
-        $this->container['incoming_characters'] = isset($data['incoming_characters']) ? $data['incoming_characters'] : null;
+        $this->container['service_accounts'] = $data['service_accounts'] ?? null;
+        $this->container['id'] = $data['id'] ?? null;
+        $this->container['name'] = $data['name'] ?? null;
+        $this->container['status'] = $data['status'] ?? null;
+        $this->container['roles'] = $data['roles'] ?? null;
+        $this->container['characters'] = $data['characters'] ?? null;
+        $this->container['groups'] = $data['groups'] ?? null;
+        $this->container['manager_groups'] = $data['manager_groups'] ?? null;
+        $this->container['manager_apps'] = $data['manager_apps'] ?? null;
+        $this->container['removed_characters'] = $data['removed_characters'] ?? null;
+        $this->container['incoming_characters'] = $data['incoming_characters'] ?? null;
     }
 
     /**
@@ -267,7 +274,8 @@ class Player implements ModelInterface, ArrayAccess
         $allowedValues = $this->getStatusAllowableValues();
         if (!is_null($this->container['status']) && !in_array($this->container['status'], $allowedValues, true)) {
             $invalidProperties[] = sprintf(
-                "invalid value for 'status', must be one of '%s'",
+                "invalid value '%s' for 'status', must be one of '%s'",
+                $this->container['status'],
                 implode("', '", $allowedValues)
             );
         }
@@ -288,6 +296,30 @@ class Player implements ModelInterface, ArrayAccess
 
 
     /**
+     * Gets service_accounts
+     *
+     * @return \Brave\NeucoreApi\Model\ServiceAccount[]|null
+     */
+    public function getServiceAccounts()
+    {
+        return $this->container['service_accounts'];
+    }
+
+    /**
+     * Sets service_accounts
+     *
+     * @param \Brave\NeucoreApi\Model\ServiceAccount[]|null $service_accounts External service accounts (API: not included by default)
+     *
+     * @return self
+     */
+    public function setServiceAccounts($service_accounts)
+    {
+        $this->container['service_accounts'] = $service_accounts;
+
+        return $this;
+    }
+
+    /**
      * Gets id
      *
      * @return int
@@ -302,7 +334,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param int $id id
      *
-     * @return $this
+     * @return self
      */
     public function setId($id)
     {
@@ -326,7 +358,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param string $name A name for the player.  This is the EVE character name of the current main character or of the last main character if there is currently none.
      *
-     * @return $this
+     * @return self
      */
     public function setName($name)
     {
@@ -350,7 +382,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param string|null $status Player account status.
      *
-     * @return $this
+     * @return self
      */
     public function setStatus($status)
     {
@@ -358,7 +390,8 @@ class Player implements ModelInterface, ArrayAccess
         if (!is_null($status) && !in_array($status, $allowedValues, true)) {
             throw new \InvalidArgumentException(
                 sprintf(
-                    "Invalid value for 'status', must be one of '%s'",
+                    "Invalid value '%s' for 'status', must be one of '%s'",
+                    $status,
                     implode("', '", $allowedValues)
                 )
             );
@@ -383,7 +416,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Role[]|null $roles Roles for authorization.
      *
-     * @return $this
+     * @return self
      */
     public function setRoles($roles)
     {
@@ -407,7 +440,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Character[]|null $characters characters
      *
-     * @return $this
+     * @return self
      */
     public function setCharacters($characters)
     {
@@ -431,7 +464,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Group[]|null $groups Group membership.
      *
-     * @return $this
+     * @return self
      */
     public function setGroups($groups)
     {
@@ -455,7 +488,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Group[]|null $manager_groups Manager of groups.
      *
-     * @return $this
+     * @return self
      */
     public function setManagerGroups($manager_groups)
     {
@@ -479,7 +512,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\App[]|null $manager_apps Manager of apps.
      *
-     * @return $this
+     * @return self
      */
     public function setManagerApps($manager_apps)
     {
@@ -503,7 +536,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\RemovedCharacter[]|null $removed_characters Characters that were removed from a player (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setRemovedCharacters($removed_characters)
     {
@@ -527,7 +560,7 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\RemovedCharacter[]|null $incoming_characters Characters that were moved from another player account to this account (API: not included by default).
      *
-     * @return $this
+     * @return self
      */
     public function setIncomingCharacters($incoming_characters)
     {
@@ -552,18 +585,18 @@ class Player implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -586,6 +619,18 @@ class Player implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
