@@ -2,7 +2,7 @@
 /**
  * RemovedCharacter
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -39,10 +38,13 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @package  Brave\NeucoreApi
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
+ * @implements \ArrayAccess<TKey, TValue>
+ * @template TKey int|null
+ * @template TValue mixed|null
  */
-class RemovedCharacter implements ModelInterface, ArrayAccess
+class RemovedCharacter implements ModelInterface, ArrayAccess, \JsonSerializable
 {
-    const DISCRIMINATOR = null;
+    public const DISCRIMINATOR = null;
 
     /**
       * The original name of the model.
@@ -71,6 +73,8 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
       * Array of property to format mappings. Used for (de)serialization
       *
       * @var string[]
+      * @phpstan-var array<string, string|null>
+      * @psalm-var array<string, string|null>
       */
     protected static $openAPIFormats = [
         'new_player_id' => null,
@@ -197,9 +201,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
     const REASON_DELETED_MANUALLY = 'deleted-manually';
     const REASON_DELETED_BIOMASSED = 'deleted-biomassed';
     const REASON_DELETED_OWNER_CHANGED = 'deleted-owner-changed';
-    
 
-    
     /**
      * Gets allowable values of the enum
      *
@@ -214,7 +216,6 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
             self::REASON_DELETED_OWNER_CHANGED,
         ];
     }
-    
 
     /**
      * Associative array for storing property values
@@ -231,14 +232,14 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      */
     public function __construct(array $data = null)
     {
-        $this->container['new_player_id'] = isset($data['new_player_id']) ? $data['new_player_id'] : null;
-        $this->container['new_player_name'] = isset($data['new_player_name']) ? $data['new_player_name'] : null;
-        $this->container['player'] = isset($data['player']) ? $data['player'] : null;
-        $this->container['character_id'] = isset($data['character_id']) ? $data['character_id'] : null;
-        $this->container['character_name'] = isset($data['character_name']) ? $data['character_name'] : null;
-        $this->container['removed_date'] = isset($data['removed_date']) ? $data['removed_date'] : null;
-        $this->container['reason'] = isset($data['reason']) ? $data['reason'] : null;
-        $this->container['deleted_by'] = isset($data['deleted_by']) ? $data['deleted_by'] : null;
+        $this->container['new_player_id'] = $data['new_player_id'] ?? null;
+        $this->container['new_player_name'] = $data['new_player_name'] ?? null;
+        $this->container['player'] = $data['player'] ?? null;
+        $this->container['character_id'] = $data['character_id'] ?? null;
+        $this->container['character_name'] = $data['character_name'] ?? null;
+        $this->container['removed_date'] = $data['removed_date'] ?? null;
+        $this->container['reason'] = $data['reason'] ?? null;
+        $this->container['deleted_by'] = $data['deleted_by'] ?? null;
     }
 
     /**
@@ -265,7 +266,8 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
         $allowedValues = $this->getReasonAllowableValues();
         if (!is_null($this->container['reason']) && !in_array($this->container['reason'], $allowedValues, true)) {
             $invalidProperties[] = sprintf(
-                "invalid value for 'reason', must be one of '%s'",
+                "invalid value '%s' for 'reason', must be one of '%s'",
+                $this->container['reason'],
                 implode("', '", $allowedValues)
             );
         }
@@ -300,7 +302,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param int|null $new_player_id new_player_id
      *
-     * @return $this
+     * @return self
      */
     public function setNewPlayerId($new_player_id)
     {
@@ -324,7 +326,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param string|null $new_player_name new_player_name
      *
-     * @return $this
+     * @return self
      */
     public function setNewPlayerName($new_player_name)
     {
@@ -348,7 +350,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Player|null $player player
      *
-     * @return $this
+     * @return self
      */
     public function setPlayer($player)
     {
@@ -372,7 +374,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param int $character_id EVE character ID.
      *
-     * @return $this
+     * @return self
      */
     public function setCharacterId($character_id)
     {
@@ -396,7 +398,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param string $character_name EVE character name.
      *
-     * @return $this
+     * @return self
      */
     public function setCharacterName($character_name)
     {
@@ -420,7 +422,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param \DateTime $removed_date Date of removal.
      *
-     * @return $this
+     * @return self
      */
     public function setRemovedDate($removed_date)
     {
@@ -444,7 +446,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param string $reason How it was removed (deleted or moved to another account).
      *
-     * @return $this
+     * @return self
      */
     public function setReason($reason)
     {
@@ -452,7 +454,8 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
         if (!in_array($reason, $allowedValues, true)) {
             throw new \InvalidArgumentException(
                 sprintf(
-                    "Invalid value for 'reason', must be one of '%s'",
+                    "Invalid value '%s' for 'reason', must be one of '%s'",
+                    $reason,
                     implode("', '", $allowedValues)
                 )
             );
@@ -477,7 +480,7 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param \Brave\NeucoreApi\Model\Player|null $deleted_by deleted_by
      *
-     * @return $this
+     * @return self
      */
     public function setDeletedBy($deleted_by)
     {
@@ -502,18 +505,18 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
      *
      * @param integer $offset Offset
      *
-     * @return mixed
+     * @return mixed|null
      */
     public function offsetGet($offset)
     {
-        return isset($this->container[$offset]) ? $this->container[$offset] : null;
+        return $this->container[$offset] ?? null;
     }
 
     /**
      * Sets value based on offset.
      *
-     * @param integer $offset Offset
-     * @param mixed   $value  Value to be set
+     * @param int|null $offset Offset
+     * @param mixed    $value  Value to be set
      *
      * @return void
      */
@@ -536,6 +539,18 @@ class RemovedCharacter implements ModelInterface, ArrayAccess
     public function offsetUnset($offset)
     {
         unset($this->container[$offset]);
+    }
+
+    /**
+     * Serializes the object to a value that can be serialized natively by json_encode().
+     * @link https://www.php.net/manual/en/jsonserializable.jsonserialize.php
+     *
+     * @return mixed Returns data which can be serialized by json_encode(), which is a value
+     * of any type other than a resource.
+     */
+    public function jsonSerialize()
+    {
+       return ObjectSerializer::sanitizeForSerialization($this);
     }
 
     /**
