@@ -2,7 +2,7 @@
 /**
  * Role
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -56,6 +55,8 @@ class Role
     const GROUP_MANAGER = 'group-manager';
     const APP_ADMIN = 'app-admin';
     const APP_MANAGER = 'app-manager';
+    const SERVICE_ADMIN = 'service-admin';
+    const STATISTICS = 'statistics';
     const ESI = 'esi';
     const SETTINGS = 'settings';
     const TRACKING = 'tracking';
@@ -84,6 +85,8 @@ class Role
             self::GROUP_MANAGER,
             self::APP_ADMIN,
             self::APP_MANAGER,
+            self::SERVICE_ADMIN,
+            self::STATISTICS,
             self::ESI,
             self::SETTINGS,
             self::TRACKING,
