@@ -1,7 +1,7 @@
 <?php
 /**
  * ApplicationApiTest
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -14,10 +14,9 @@
  *
  * Client library of Neucore API
  *
- * OpenAPI spec version: 0.7.1
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 3.3.4
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -26,11 +25,12 @@
  * Please update the test case below to test the endpoint.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Api;
 
 use \Brave\NeucoreApi\Configuration;
 use \Brave\NeucoreApi\ApiException;
 use \Brave\NeucoreApi\ObjectSerializer;
+use PHPUnit\Framework\TestCase;
 
 /**
  * ApplicationApiTest Class Doc Comment
@@ -40,194 +40,34 @@ use \Brave\NeucoreApi\ObjectSerializer;
  * @author   OpenAPI Generator team
  * @link     https://openapi-generator.tech
  */
-class ApplicationApiTest extends \PHPUnit_Framework_TestCase
+class ApplicationApiTest extends TestCase
 {
 
     /**
      * Setup before running any test cases
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
-    {
-    }
-
-    /**
-     * Test case for allianceGroupsBulkV1
-     *
-     * Return groups of multiple alliances..
-     *
-     */
-    public function testAllianceGroupsBulkV1()
-    {
-    }
-
-    /**
-     * Test case for allianceGroupsV1
-     *
-     * Return groups of the alliance..
-     *
-     */
-    public function testAllianceGroupsV1()
-    {
-    }
-
-    /**
-     * Test case for allianceGroupsV2
-     *
-     * Return groups of the alliance..
-     *
-     */
-    public function testAllianceGroupsV2()
-    {
-    }
-
-    /**
-     * Test case for charactersV1
-     *
-     * Return all characters of the player account to which the character ID belongs..
-     *
-     */
-    public function testCharactersV1()
-    {
-    }
-
-    /**
-     * Test case for corpGroupsBulkV1
-     *
-     * Return groups of multiple corporations..
-     *
-     */
-    public function testCorpGroupsBulkV1()
-    {
-    }
-
-    /**
-     * Test case for corpGroupsV1
-     *
-     * Return groups of the corporation..
-     *
-     */
-    public function testCorpGroupsV1()
-    {
-    }
-
-    /**
-     * Test case for corpGroupsV2
-     *
-     * Return groups of the corporation..
-     *
-     */
-    public function testCorpGroupsV2()
-    {
-    }
-
-    /**
-     * Test case for esiPostV1
-     *
-     * Makes an ESI POST request on behalf on an EVE character and returns the result..
-     *
-     */
-    public function testEsiPostV1()
-    {
-    }
-
-    /**
-     * Test case for esiV1
-     *
-     * Makes an ESI GET request on behalf on an EVE character and returns the result..
-     *
-     */
-    public function testEsiV1()
-    {
-    }
-
-    /**
-     * Test case for groupsBulkV1
-     *
-     * Return groups of multiple players, identified by one of their character IDs..
-     *
-     */
-    public function testGroupsBulkV1()
-    {
-    }
-
-    /**
-     * Test case for groupsV1
-     *
-     * Return groups of the character's player account..
-     *
-     */
-    public function testGroupsV1()
-    {
-    }
-
-    /**
-     * Test case for groupsV2
-     *
-     * Return groups of the character's player account..
-     *
-     */
-    public function testGroupsV2()
-    {
-    }
-
-    /**
-     * Test case for groupsWithFallbackV1
-     *
-     * Returns groups from the character's account, if available, or the corporation and alliance..
-     *
-     */
-    public function testGroupsWithFallbackV1()
-    {
-    }
-
-    /**
-     * Test case for mainV1
-     *
-     * Returns the main character of the player account to which the character ID belongs..
-     *
-     */
-    public function testMainV1()
-    {
-    }
-
-    /**
-     * Test case for mainV2
-     *
-     * Return the main character of the player account to which the character ID belongs..
-     *
-     */
-    public function testMainV2()
-    {
-    }
-
-    /**
-     * Test case for memberTrackingV1
-     *
-     * Return corporation member tracking data..
-     *
-     */
-    public function testMemberTrackingV1()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -239,5 +79,7 @@ class ApplicationApiTest extends \PHPUnit_Framework_TestCase
      */
     public function testShowV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
