@@ -1,7 +1,7 @@
 <?php
 /**
  * ApplicationCharactersApiTest
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -14,10 +14,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -26,7 +25,7 @@
  * Please update the test case below to test the endpoint.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Api;
 
 use \Brave\NeucoreApi\Configuration;
 use \Brave\NeucoreApi\ApiException;
@@ -47,28 +46,28 @@ class ApplicationCharactersApiTest extends TestCase
     /**
      * Setup before running any test cases
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -80,6 +79,20 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testCharacterListV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test case for charactersBulkV1
+     *
+     * Returns all characters from multiple player accounts identified by character IDs..
+     *
+     */
+    public function testCharactersBulkV1()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -90,6 +103,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -100,6 +115,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testCorporationCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -110,6 +127,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testCorporationPlayersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -120,6 +139,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testIncomingCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -130,6 +151,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testMainV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -140,6 +163,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testMainV2()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -150,6 +175,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testPlayerCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -160,6 +187,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testPlayerV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -170,6 +199,8 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testPlayerWithCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -180,5 +211,7 @@ class ApplicationCharactersApiTest extends TestCase
      */
     public function testRemovedCharactersV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
