@@ -1,7 +1,7 @@
 <?php
 /**
  * ApplicationTrackingApiTest
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -14,10 +14,9 @@
  *
  * Client library of Neucore API
  *
- * The version of the OpenAPI document: 1.14.0
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 4.3.1
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -26,7 +25,7 @@
  * Please update the test case below to test the endpoint.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Api;
 
 use \Brave\NeucoreApi\Configuration;
 use \Brave\NeucoreApi\ApiException;
@@ -47,28 +46,28 @@ class ApplicationTrackingApiTest extends TestCase
     /**
      * Setup before running any test cases
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -80,5 +79,7 @@ class ApplicationTrackingApiTest extends TestCase
      */
     public function testMemberTrackingV1()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
