@@ -2,7 +2,7 @@
 /**
  * CharacterTest
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * OpenAPI spec version: 0.7.1
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 3.3.4
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -27,7 +26,9 @@
  * Please update the test case below to test the model.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Model;
+
+use PHPUnit\Framework\TestCase;
 
 /**
  * CharacterTest Class Doc Comment
@@ -38,34 +39,34 @@ namespace Brave\NeucoreApi;
  * @author      OpenAPI Generator team
  * @link        https://openapi-generator.tech
  */
-class CharacterTest extends \PHPUnit_Framework_TestCase
+class CharacterTest extends TestCase
 {
 
     /**
      * Setup before running any test case
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -74,6 +75,8 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testCharacter()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -81,6 +84,8 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyId()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -88,6 +93,8 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyName()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -95,6 +102,8 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyMain()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -102,6 +111,26 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyValidToken()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "valid_token_time"
+     */
+    public function testPropertyValidTokenTime()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "created"
+     */
+    public function testPropertyCreated()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -109,6 +138,8 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyLastUpdate()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -116,5 +147,16 @@ class CharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyCorporation()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "character_name_changes"
+     */
+    public function testPropertyCharacterNameChanges()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
