@@ -2,7 +2,7 @@
 /**
  * CorporationMemberTest
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * OpenAPI spec version: 0.7.1
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 3.3.4
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -27,7 +26,9 @@
  * Please update the test case below to test the model.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Model;
+
+use PHPUnit\Framework\TestCase;
 
 /**
  * CorporationMemberTest Class Doc Comment
@@ -38,34 +39,34 @@ namespace Brave\NeucoreApi;
  * @author      OpenAPI Generator team
  * @link        https://openapi-generator.tech
  */
-class CorporationMemberTest extends \PHPUnit_Framework_TestCase
+class CorporationMemberTest extends TestCase
 {
 
     /**
      * Setup before running any test case
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -74,6 +75,8 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testCorporationMember()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -81,6 +84,8 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyPlayer()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -88,6 +93,8 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyId()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -95,13 +102,17 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyName()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
-     * Test attribute "location_id"
+     * Test attribute "location"
      */
-    public function testPropertyLocationId()
+    public function testPropertyLocation()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -109,6 +120,8 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyLogoffDate()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -116,13 +129,17 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyLogonDate()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
-     * Test attribute "ship_type_id"
+     * Test attribute "ship_type"
      */
-    public function testPropertyShipTypeId()
+    public function testPropertyShipType()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -130,6 +147,8 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyStartDate()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -137,5 +156,34 @@ class CorporationMemberTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyCharacter()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "missing_character_mail_sent_date"
+     */
+    public function testPropertyMissingCharacterMailSentDate()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "missing_character_mail_sent_result"
+     */
+    public function testPropertyMissingCharacterMailSentResult()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "missing_character_mail_sent_number"
+     */
+    public function testPropertyMissingCharacterMailSentNumber()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
