@@ -2,7 +2,7 @@
 /**
  * PlayerTest
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * OpenAPI spec version: 0.7.1
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 3.3.4
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -27,7 +26,9 @@
  * Please update the test case below to test the model.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Model;
+
+use PHPUnit\Framework\TestCase;
 
 /**
  * PlayerTest Class Doc Comment
@@ -38,34 +39,34 @@ namespace Brave\NeucoreApi;
  * @author      OpenAPI Generator team
  * @link        https://openapi-generator.tech
  */
-class PlayerTest extends \PHPUnit_Framework_TestCase
+class PlayerTest extends TestCase
 {
 
     /**
      * Setup before running any test case
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -74,6 +75,17 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPlayer()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "service_accounts"
+     */
+    public function testPropertyServiceAccounts()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -81,6 +93,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyId()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -88,6 +102,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyName()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -95,6 +111,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyStatus()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -102,6 +120,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyRoles()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -109,6 +129,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyCharacters()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -116,6 +138,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyGroups()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -123,6 +147,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyManagerGroups()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -130,6 +156,8 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyManagerApps()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -137,5 +165,16 @@ class PlayerTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyRemovedCharacters()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "incoming_characters"
+     */
+    public function testPropertyIncomingCharacters()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
