@@ -2,7 +2,7 @@
 /**
  * RemovedCharacterTest
  *
- * PHP version 5
+ * PHP version 7.2
  *
  * @category Class
  * @package  Brave\NeucoreApi
@@ -15,10 +15,9 @@
  *
  * Client library of Neucore API
  *
- * OpenAPI spec version: 0.7.1
- * 
+ * The version of the OpenAPI document: 1.21.0
  * Generated by: https://openapi-generator.tech
- * OpenAPI Generator version: 3.3.4
+ * OpenAPI Generator version: 5.1.1
  */
 
 /**
@@ -27,7 +26,9 @@
  * Please update the test case below to test the model.
  */
 
-namespace Brave\NeucoreApi;
+namespace Brave\NeucoreApi\Test\Model;
+
+use PHPUnit\Framework\TestCase;
 
 /**
  * RemovedCharacterTest Class Doc Comment
@@ -38,34 +39,34 @@ namespace Brave\NeucoreApi;
  * @author      OpenAPI Generator team
  * @link        https://openapi-generator.tech
  */
-class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
+class RemovedCharacterTest extends TestCase
 {
 
     /**
      * Setup before running any test case
      */
-    public static function setUpBeforeClass()
+    public static function setUpBeforeClass(): void
     {
     }
 
     /**
      * Setup before running each test case
      */
-    public function setUp()
+    public function setUp(): void
     {
     }
 
     /**
      * Clean up after running each test case
      */
-    public function tearDown()
+    public function tearDown(): void
     {
     }
 
     /**
      * Clean up after running all test cases
      */
-    public static function tearDownAfterClass()
+    public static function tearDownAfterClass(): void
     {
     }
 
@@ -74,6 +75,8 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testRemovedCharacter()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -81,6 +84,8 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyNewPlayerId()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -88,6 +93,17 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyNewPlayerName()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "player"
+     */
+    public function testPropertyPlayer()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -95,6 +111,8 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyCharacterId()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -102,6 +120,8 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyCharacterName()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -109,6 +129,8 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyRemovedDate()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 
     /**
@@ -116,5 +138,16 @@ class RemovedCharacterTest extends \PHPUnit_Framework_TestCase
      */
     public function testPropertyReason()
     {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
+    }
+
+    /**
+     * Test attribute "deleted_by"
+     */
+    public function testPropertyDeletedBy()
+    {
+        // TODO: implement
+        $this->markTestIncomplete('Not implemented');
     }
 }
